cleanup module imports
